移除了不完全的代码匹配内容
只看标题匹配的话还有一种情况，就是标题中的单词的定语在现在的so社区被删除导致不能匹配，实际上可以稍微修改一下进行匹配，但是不能保证是否完全是同一个题目了。
