+++ elastic_tool.py
@@ -1,6 +1,9 @@
 import ipdb
 import xmltodict
 import html
+import json
+import linecache
+
 from elasticsearch import Elasticsearch, helpers
 from logging import config,getLogger
 
@@ -70,10 +73,47 @@ def query_index(es, field, support_str):
     logger.info(top_three_simple)
     return top_three_simple
 
+def query_file(es,read_file_path,write_file_path):
+    """
+    查询指定文件的所有标题，并把排名前三的结果写入json
+    """
+    with open(read_file_path, mode='r', encoding='utf-8') as fr:
+        with open(write_file_path,mode="w",encoding="utf-8") as fw:
+            for line in fr:
+                line_dict = {} 
+                # line_dict["title"] = line
+                line_dict["code"] = line
+                # line_dict["top_three_simple"] = query_index(es,"title",line)
+                line_dict["top_three_simple"] = query_index(es,"body",line)
+
+                line_json = json.dumps(line_dict)
+
+                fw.write(line_json + '\n') 
+            fw.close()
+def query_jsonfile(es,read_file_path_json,read_file_path_code,write_file_path):
+    '''
+    读取json文件，并进行代码段匹配，把排名前三的写入json
+    '''
+    with open(read_file_path_json, mode='r', encoding='utf-8') as fr_json:
+        with open(write_file_path,mode="w",encoding="utf-8") as fw:
+            for line in fr_json:
+                line_dict={} #用于生成json文件
+                load_json = json.loads(line)
+                line_index=load_json["linedex"]
+                line_dict["code"] = linecache.getline(read_file_path_code, line_index)
+                line_dict["top_three_simple"] = query_index(es,"body",line_dict["code"])
+                line_dict["line_index"] = line_index
+
+                line_json = json.dumps(line_dict)
+                fw.write(line_json+"\n")
+            fw.close()
+
 if __name__ == '__main__':
     es = get_connection()
     # initialize(es)
-    query_index(es, 'title', 'how to format python string based on byte length')
+    query_file(es,"/home/zzm/sdb2_zzm/Code2Que-master/Code2Que-data/pydata/src-train.txt","/home/zzm/sdb2_zzm/Code2Que-master/Code2Que-data/pydata/src-train.json")
+    # query_jsonfile(es,"notmatch-pytgt-train.json","/home/zzm/sdb2_zzm/Code2Que-master/Code2Que-data/pydata/src-train.txt","notmatch-pysrc-train.json")
+    # query_index(es, 'title', 'how to format python string based on byte length ?')
     # for_line_in("data/xml-data/with_python_tag_all.xml", TOTAL_SIZE, BULK_SIZE, upload, es)
